+++ replacer-backend/src/main/java/es/bvalero/replacer/repository/PageRepository.java
@@ -23,13 +23,13 @@ public interface PageRepository {
     Collection<Integer> findPageIdsToReview(WikipediaLanguage lang, int numResult);
 
     /** Count the number of pages to review */
-    long countPagesToReview(WikipediaLanguage lang);
+    int countPagesToReview(WikipediaLanguage lang);
 
     /** Find a random batch of pages to review for a given type and return the IDs */
     Collection<Integer> findPageIdsToReviewByType(WikipediaLanguage lang, ReplacementType type, int numResult);
 
     /** Count the number of pages to review by type */
-    long countPagesToReviewByType(WikipediaLanguage lang, ReplacementType type);
+    int countPagesToReviewByType(WikipediaLanguage lang, ReplacementType type);
 
     /** Find the pages to review by the given type and return the titles */
     Collection<String> findPageTitlesToReviewByType(WikipediaLanguage lang, ReplacementType type);
